pull out eap decorator, simplfy test
