+++ tests/conftest.py
@@ -73,7 +73,9 @@ def create_databases() -> None:
 
 def pytest_collection_modifyitems(items: Sequence[Any]) -> None:
     for item in items:
-        if item.get_closest_marker("clickhouse_db"):
+        if item.get_closest_marker("eap"):
+            item.fixturenames.append("eap")
+        elif item.get_closest_marker("clickhouse_db"):
             item.fixturenames.append("clickhouse_db")
         elif item.get_closest_marker("custom_clickhouse_db"):
             item.fixturenames.append("custom_clickhouse_db")
@@ -272,6 +274,31 @@ def clickhouse_db(
                     )
                     connection.execute(create_table_query)
                 applied_nodes.add((node.host_name, node.port, table_name))
+        yield
+    finally:
+        _clear_db()
+
+
+@pytest.fixture
+def eap(request: pytest.FixtureRequest, create_databases: None) -> Generator[None, None, None]:
+    """
+    A custom ClickHouse fixture that only runs EAP (Events Analytics Platform) migrations and Outcomes migrations (for storage routing).
+    This is much faster than running all migrations for tests that only need EAP tables.
+
+    Use this with @pytest.mark.eap marker.
+    """
+    if not request.node.get_closest_marker("eap"):
+        pytest.fail("Need to use eap marker if eap fixture is used")
+
+    from snuba.migrations.groups import MigrationGroup
+    from snuba.migrations.runner import Runner
+
+    try:
+        reset_dataset_factory()
+        # Run only SYSTEM migrations (required for migrations table) and EAP migrations
+        runner = Runner()
+        runner.run_all(group=MigrationGroup.EVENTS_ANALYTICS_PLATFORM, force=True)
+        runner.run_all(group=MigrationGroup.OUTCOMES, force=True)
         yield
     finally:
         _clear_db()
